Added voice chats Fixed errors

+++ src/main/java/ru/app/voicechat/controllers/AudioController.java
@@ -5,10 +5,13 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import ru.app.voicechat.models.Recording;
+import ru.app.voicechat.models.Room;
+import ru.app.voicechat.models.User;
 import ru.app.voicechat.services.AudioService;
 import ru.app.voicechat.services.UserService;
 
 import java.util.Comparator;
+import java.util.stream.Collectors;
 
 @Controller
 public class AudioController {
@@ -34,8 +37,7 @@ public class AudioController {
         if (recording.isPresent() &&
                 (recording.get().getOwner().equals(userService.getUser()) ||
                         (!recording.get().getOwner().equals(userService.getUser()) &&
-                                !recording.get().getPrivateToOwner())))
-        {
+                                !recording.get().getPrivateToOwner()))) {
             model.addAttribute("recording", recording.get());
             return "curr_recording";
         }
@@ -62,8 +64,52 @@ public class AudioController {
     }
 
     @PostMapping("/rec/delete")
-    public String deleteProduct(@RequestParam int id) {
-        audioService.removeRecording(Integer.toUnsignedLong(id));
-        return "redirect:/recordings";
+    public String deleteProduct(@RequestParam int id, Model model) {
+        var recording = audioService.getRecording(Integer.toUnsignedLong(id));
+        if (recording.isPresent())
+            if (recording.get().getOwner().equals(userService.getUser())) {
+                audioService.removeRecording(Integer.toUnsignedLong(id));
+                return "redirect:/recordings";
+            } else {
+                model.addAttribute("reason", "Это запись с id " + id + " принадлежит не вам!");
+                return "error";
+            }
+        else {
+            model.addAttribute("reason", "Не найдена запись с id " + id);
+            return "error";
+        }
+    }
+
+    @GetMapping("/voicechats")
+    public String getVoiceChats(Model model) {
+        model.addAttribute("voice_chats", audioService.getRooms());
+        return "voice_chats";
+    }
+
+    @GetMapping("/vc/new")
+    public String getNewVC(Model model) {
+        model.addAttribute("new_voice_chat", new Room());
+        return "new_voice_chat";
+    }
+
+    @PostMapping("/vc/new")
+    public String setNewVC(@ModelAttribute("voice_chat") Room new_room) {
+        audioService.saveRoom(new_room);
+        return "redirect:/voicechats";
+    }
+
+    @GetMapping("/vc/{id}")
+    public String getVC(@PathVariable int id, Model model) {
+        var room = audioService.getRoom(Integer.toUnsignedLong(id));
+        if (room.isPresent()) {
+            model.addAttribute("room", room.get());
+            model.addAttribute("userList",
+                    room.get().getUserList().stream().map(User::getUsername).collect(Collectors.joining(", ")));
+            model.addAttribute("user", userService.getUser());
+            return "curr_voice_chat";
+        } else {
+            model.addAttribute("reason", "Не найден чат с id " + id);
+            return "error";
+        }
     }
 }
